refactor: move hover update to stage component
